+++ NEWS.md
@@ -14,11 +14,13 @@
 * Fixed bug in geom_pointpath() (#15)
 * Fixed scale_(x/y)_dendrogram so that labels are passed from scale (#17)
 * geom_pointpath() now curves with nonlinear coordinates (#15).
-* Added stat_funxy(), stat_centroid() and stat_midpoint() (#16)
+* Added stat_funxy(), stat_centroid() and stat_midpoint() (#16).
 * Added facet_wrap2() with few extensions.
 * Added facet_nested_wrap() for merging strips (#19)
 * Alternative specification of facet scales in `facetted_pos_scales` through 
   formulas (#25).
+* Added option to discard dendrogram labels (#23).
+* Added coloured text legend (stringlegend, #31).
 
 # ggh4x 0.1.0